Some renaming and cleanup
